prepared for uint32_t sample value type

+++ src/avr-common/utils/uc-core/synchronization/BasicCalculationTypes.h
@@ -15,7 +15,8 @@
 #  define IndexType uint8_t
 #else
 typedef uint32_t CumulationType;
-typedef uint16_t SampleValueType;
+//typedef uint16_t SampleValueType;
+typedef uint32_t SampleValueType;
 //typedef float CalculationType;
 typedef double CalculationType;
 typedef uint8_t IndexType;
